Create aggregator for transaction page
