+++ lib/aggregators/home-info.ts
@@ -35,7 +35,7 @@ class HomeInfo extends Aggregator<HomeInfoResult> {
   async setter() {
     const [counts, nameOperations] = await Promise.all([
       NameCounts.fetch(),
-      NameOperations.setter({page: 0}),
+      NameOperations.fetch({page: 0}),
     ]);
 
     const startCount = counts.total - nameOperations.length;
